+++ include/Player/Characters/Character.h
@@ -29,17 +29,17 @@ protected:
 protected:
   // Generic Class Variables
   string className;
-  short int health;
-  short unsigned int xp = 0;
-  short unsigned int level = 1;
-  short unsigned int meleeWeapon = 1, magicWeapon = 1, rangedWeapon = 1;
+  int health;
+  int xp = 0;
+  int level = 1;
+  int meleeWeapon = 1, magicWeapon = 1, rangedWeapon = 1;
 
   // Class Specific Variables
-  short unsigned int maxHealth = 0;
-  short unsigned int meleeCounter = 0, magicCounter = 0, rangedCounter = 0;
-  short unsigned int meleeUpgradeTier = 0, magicUpgradeTier = 0, rangedUpgradeTier = 0;
-  short unsigned int meleeLevel = 0, rangedLevel = 0, magicLevel = 0;
-  short unsigned int xpRequiredPerLevel = 100;
+  int maxHealth = 0;
+  int meleeCounter = 0, magicCounter = 0, rangedCounter = 0;
+  int meleeUpgradeTier = 0, magicUpgradeTier = 0, rangedUpgradeTier = 0;
+  int meleeLevel = 0, rangedLevel = 0, magicLevel = 0;
+  int xpRequiredPerLevel = 100;
   double critLevel = 1.0;
   // Map of skill names with array for indicating skill level - key is ability type [ Melee, Magic, or Ranged ]
   unordered_map<string, string [3]> skillNames;
@@ -52,29 +52,29 @@ protected:
 public:
   // Generic Class Accessors
   string getClassName() { return className; }
-  short int getHealth() { return health; }
-  short unsigned int getXp() { return xp; }
-  short unsigned int getLevel() { return level; }
-  short unsigned int getWeaponLevel(const string &weaponType);
+  int getHealth() { return health; }
+  int getXp() { return xp; }
+  int getLevel() { return level; }
+  int getWeaponLevel(const string &weaponType);
 
   // Class Specific Accessors
-  short unsigned int getMaxHealth() { return maxHealth; }
-  short unsigned int getSkillUpgradeTier(const string &skillType);
-  short unsigned int getSkillLevel(const string &skillType);
+  int getMaxHealth() { return maxHealth; }
+  int getSkillUpgradeTier(const string &skillType);
+  int getSkillLevel(const string &skillType);
   double getCritLevel() { return critLevel; }
   // Retrieves value from skillNames
   string getSkillName(const string &skillType);
 
   // Generic Class Mutators
-  void adjustHealth(short int healthAdjust) { health += healthAdjust; }
-  void addXp(short unsigned int xpAdjust);
+  void adjustHealth(int healthAdjust) { health += healthAdjust; }
+  void addXp(int xpAdjust);
   void addLevel();
   void upgradeWeapon(const string &weaponType, const string &upgradeName);
 
   // Class Specific Mutators
   void useSkill(const string &skillType);
   void upgradeSkillName(const string &skillType);
-  void adjustMaxHealth(short unsigned int maxHealthAdjust) { maxHealth += maxHealthAdjust; }
+  void adjustMaxHealth(int maxHealthAdjust) { maxHealth += maxHealthAdjust; }
 };
 
 #endif // DASTARDLY_DUNGEON_DWELLERS_CHARACTER_H
